Remove weird legacy unicode stuff
First of all we've decided to no longer support 2.6 so that's not a consideration anymore. The rest should just work like any other class really. I'm also deleting the test because it assumes that any character which it doesn't understand will be replaced by a '?'. This is no longer the case, nor should it be necessary.